# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CalibrateRequest": {
        "properties": {
          "distance_mm": {
            "exclusiveMinimum": 0.0,
            "title": "Distance Mm",
            "type": "number"
          },
          "points": {
            "items": {
              "items": {
                "type": "number"
              },
              "maxItems": 2,
              "minItems": 2,
              "type": "array"
            },
            "maxItems": 2,
            "minItems": 2,
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "points",
          "distance_mm"
        ],
        "title": "CalibrateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocationOut": {
        "properties": {
          "grid_index": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grid Index"
          },
          "grid_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grid Reference"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "view_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Id"
          },
          "view_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Kind"
          },
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          }
        },
        "required": [
          "view_id",
          "view_kind",
          "grid_reference",
          "grid_index",
          "level",
          "zone"
        ],
        "title": "LocationOut",
        "type": "object"
      },
      "MeasureOut": {
        "properties": {
          "denominator": {
            "title": "Denominator",
            "type": "number"
          },
          "length_mm": {
            "title": "Length Mm",
            "type": "number"
          },
          "scale_status": {
            "title": "Scale Status",
            "type": "string"
          }
        },
        "required": [
          "length_mm",
          "denominator",
          "scale_status"
        ],
        "title": "MeasureOut",
        "type": "object"
      },
      "MeasureRequest": {
        "properties": {
          "points": {
            "items": {
              "items": {
                "type": "number"
              },
              "maxItems": 2,
              "minItems": 2,
              "type": "array"
            },
            "maxItems": 10000,
            "minItems": 2,
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "points"
        ],
        "title": "MeasureRequest",
        "type": "object"
      },
...
      "SheetDetail": {
        "properties": {
          "base_height_px": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Height Px"
          },
          "base_width_px": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Width Px"
          },
          "content_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Class"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "filename": {
            "default": "",
            "title": "Filename",
            "type": "string"
          },
          "has_thumbnail": {
            "default": false,
            "title": "Has Thumbnail",
            "type": "boolean"
          },
          "height_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Mm"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "index_in_document": {
            "title": "Index In Document",
            "type": "integer"
          },
          "layout_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout Name"
          },
          "manual_takeoff_recommended": {
            "default": false,
            "title": "Manual Takeoff Recommended",
            "type": "boolean"
          },
          "max_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Level"
          },
          "quality_band": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Band"
          },
          "quality_detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Detail"
          },
          "source_ref": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "tile_source": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tile Source"
          },
          "width_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width Mm"
          }
        },
        "required": [
          "id",
          "document_id",
          "index_in_document",
          "layout_name",
          "width_mm",
          "height_mm",
          "content_class",
          "max_level",
          "base_width_px",
          "base_height_px"
        ],
        "title": "SheetDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ViewOut": {
        "properties": {
          "calibrated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibrated By"
          },
          "calibration": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration"
          },
          "denominator": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Denominator"
          },
          "extent": {
            "items": {
              "type": "number"
            },
            "title": "Extent",
            "type": "array"
          },
          "grid": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grid"
          },
          "grid_box": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grid Box"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "measurable": {
            "default": false,
            "title": "Measurable",
            "type": "boolean"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "scale_evidence": {
            "additionalProperties": true,
            "title": "Scale Evidence",
            "type": "object"
          },
          "scale_status": {
            "title": "Scale Status",
            "type": "string"
          },
          "sheet_id": {
            "format": "uuid",
            "title": "Sheet Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stated_denominator": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stated Denominator"
          },
          "stated_scale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stated Scale"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "sheet_id",
          "ordinal",
          "kind",
          "title",
          "source",
          "extent",
          "level",
          "stated_scale",
          "stated_denominator",
          "scale_status",
          "denominator",
          "scale_evidence",
          "grid",
          "grid_box",
          "calibration",
          "calibrated_by"
        ],
        "title": "ViewOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "OIDC access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FireBid SG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/bids/{bid_id}/sheets/{sheet_id}": {
      "get": {
        "operationId": "get_sheet_bids__bid_id__sheets__sheet_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "sheet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sheet Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bid Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SheetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Sheet",
        "tags": [
          "sheets"
        ]
      }
    },
    "/bids/{bid_id}/sheets/{sheet_id}/locate": {
      "get": {
        "description": "A sheet point as an estimator says it: grid reference, level and zone.",
        "operationId": "locate_point_bids__bid_id__sheets__sheet_id__locate_get",
        "parameters": [
          {
            "in": "path",
            "name": "sheet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sheet Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bid Id",
              "type": "string"
            }
          },
          {
            "description": "sheet millimetres from the left",
            "in": "query",
            "name": "x",
            "required": true,
            "schema": {
              "description": "sheet millimetres from the left",
              "title": "X",
              "type": "number"
            }
          },
          {
            "description": "sheet millimetres from the top",
            "in": "query",
            "name": "y",
            "required": true,
            "schema": {
              "description": "sheet millimetres from the top",
              "title": "Y",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Locate Point",
        "tags": [
          "views"
        ]
      }
    },
...
    "/bids/{bid_id}/sheets/{sheet_id}/views": {
      "get": {
        "operationId": "list_views_bids__bid_id__sheets__sheet_id__views_get",
        "parameters": [
          {
            "in": "path",
            "name": "sheet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sheet Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bid Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ViewOut"
                  },
                  "title": "Response List Views Bids  Bid Id  Sheets  Sheet Id  Views Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Views",
        "tags": [
          "views"
        ]
      }
    },
    "/bids/{bid_id}/sheets/{sheet_id}/views/{view_id}/calibrate": {
      "post": {
        "description": "Two sheet points and the real distance between them: the view becomes measurable.",
        "operationId": "calibrate_view_bids__bid_id__sheets__sheet_id__views__view_id__calibrate_post",
        "parameters": [
          {
            "in": "path",
            "name": "sheet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sheet Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "view_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "View Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bid Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Calibrate View",
        "tags": [
          "views"
        ]
      }
    },
    "/bids/{bid_id}/sheets/{sheet_id}/views/{view_id}/measure": {
      "post": {
        "description": "A length along sheet points, in millimetres. 409 unless the scale is measurable.",
        "operationId": "measure_on_view_bids__bid_id__sheets__sheet_id__views__view_id__measure_post",
        "parameters": [
          {
            "in": "path",
            "name": "sheet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sheet Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "view_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "View Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bid Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasureOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Measure On View",
        "tags": [
          "views"
        ]
      }
    },
...
  }
}